% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lspline.R
\name{lspline}
\alias{lspline}
\title{Basis for a piecewise linear spline}
\usage{
lspline(x, knots = NULL, marginal = FALSE, names = NULL)
}
\arguments{
\item{x}{numeric vector, the variable}

\item{knots}{numeric vector of knots positions}

\item{marginal}{logical, how to parametrize the spline, see Details}

\item{names}{character, vector of names for constructed variables}
}
\description{
Basis for a piecewise linear spline
}
\references{
Junger & Ponce de Leon (2011) "ares: Environment air
pollution epidemiology: a library for timeseries
analysis". R package version 0.7.2 retrieved from CRAN
archives.
}
\author{
Based on \code{ares::lspline} from Junger & Ponce de Leon (2011)
}
