% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinear.R
\name{plinear}
\alias{plinear}
\title{Piecewise-linear spline}
\usage{
plinear(x, ...)
}
\arguments{
\item{x}{numeric predictor}

\item{...}{numeric values of x-positions of the knots}
}
\value{
A matrix with dimensionality (length(x), number-of-knots)
}
\description{
Piecewise-linear spline
}
\examples{
set.seed(666)
d <- data.frame(
  x = 1:10,
  y = c(5:1, 2:6) + rnorm(10, 0, 0.5)
)
mod <- lm( y ~ x + plinear(x, 5), data=d)
d$f <- fitted(mod)

plot(y ~ x, data=d)
lines(f ~ x, data=d)

coef(mod)["x"] # Slope for 'x' in (-Inf, 5)
coef(mod)[3] # Slope change at knot x=5
coef(mod)["x"] + coef(mod)[3] # Slope for 'x' in (5, Inf)
}
